requirements.txt
numpy>=1.17.0
opencv-python>=4.1.0
pillow>=6.1.0